refactor: move the e2e-test into the attributes test
